seperates join room and study_rooms
