+++ client/src/rooms/chatlogic.jsx
@@ -1,0 +1,29 @@
+
+      {/* move this logic to the Formik Join Room form */}
+      {/* <div className="StudyChat">
+        {!showChat ? (
+          <div className="joinChatContainer">
+            <h3>Join Chat</h3>
+            <input
+              type="text"
+              placeholder="Name"
+              onChange={(event) => {
+                setUsername(event.target.value);
+              }}
+            />
+            <input
+              type="text"
+              placeholder="Room ID"
+              onChange={(event) => {
+                setRoom(event.target.value);
+              }}
+            />
+            <button onClick={joinRoom}>Join a Room</button>
+          </div>
+        ) : (
+
+            //Keep this part 
+            <Chat socket={socket} username={username} room={room} /> 
+            
+        )}
+      </div> */}
