+++ client/src/rooms/study-room.jsx
@@ -1,5 +1,5 @@
 import io from 'socket.io-client';
-import { useState } from 'react';
+import { useState, useEffect } from 'react';
 import {
   Chat,
   Users,
@@ -8,7 +8,7 @@ import {
   Sound
 } from "./index";
 import '../styles/app.scss';
-// import { Formik } from "formik";
+import { Formik } from "formik";
 import {
   Heading,
   Button
@@ -17,92 +17,26 @@ import * as Yup from 'yup';
 import TextField from '../components/TextField';
 import axios from 'axios';
 
-const socket = io.connect("/");
+// const socket = io.connect("/");
 
 
-export const StudyRoom = () => {
-
-  const [username, setUsername] = useState("");
-  const [room, setRoom] = useState("");
-  const [showChat, setShowChat] = useState(false);
-  const [error, setError] = useState("");
-  const [joinStatus, setJoinStatus] = useState("");
+export const StudyRoom = (props) => {
 
 
-  const joinRoom = () => {
-    if (username !== "" && room !== "") {
-      socket.emit("join_room", room);
-      setShowChat(true);
-    }
-  };
+
+  const username = props.username
+  const socket = props.socket
+  const usersList = props.usersList
+  const room = props.room
+   const users = usersList.map(user => {
+    return (<Users username = {user}/>)
+  })
+
 
 
   return (
 
     <main className='study-main'>
-
-      {/* <div className="join-room-form">
-      <Formik
-        initialValues={{
-          id: "",
-          username: "",
-        }}
-        validationSchema={Yup.object({
-          id: Yup.number()
-            .required("Room ID Required"),
-          username: Yup.string()
-          .required("Username Required")
-        })}
-        onSubmit={(values, actions) => {
-          alert(JSON.stringify(values, null, 2));
-          actions.resetForm();
-        }}
-      >
-        {formik => (
-          <Flex
-            bg="white.100"
-            align="center"
-            justify="center"
-            h="60vh"
-          >
-            <Box
-              bg="white"
-              p={6}
-              rounded="md"
-              w={80}
-            >
-              <VStack
-                as="form"
-                mx="auto"
-                spacing={5}
-                justifyContent="center"
-                onSubmit={formik.handleSubmit}
-              >
-                <Heading>Join a Study Room</Heading>
-
-                <TextField
-                  name="id"
-                  type="number"
-                  placeholder="Room ID"
-                />
-
-                <TextField
-                  name="username"
-                  type="username"
-                  placeholder="Username"
-                />
-
-                <Button type="submit" colorScheme="purple" width="full">
-                  Start Studying!
-                </Button>
-
-              </VStack>
-            </Box>
-          </Flex>
-        )}
-        </Formik>
-    </div> */}
-
 
       <div className="study-dashboard">
 
@@ -120,7 +54,8 @@ export const StudyRoom = () => {
         <div className="study-components">
           <div className="left-study-bar">
             <div className="users-component">
-              <Users username={username} />
+            <Heading size="sm">Study Buddies</Heading>
+              {users}
             </div>
           </div>
 
@@ -132,7 +67,7 @@ export const StudyRoom = () => {
 
           <div className="right-study-bar">
             <div className="chat-component">
-              <Chat socket={socket} username={username} room={room} />
+              <Chat  socket={socket} username={username} room={room}  />
             </div>
 
             <div className="notes-component">
@@ -148,34 +83,6 @@ export const StudyRoom = () => {
       </div>
 
 
-      {/* move this logic to the Formik Join Room form */}
-      {/* <div className="StudyChat">
-        {!showChat ? (
-          <div className="joinChatContainer">
-            <h3>Join Chat</h3>
-            <input
-              type="text"
-              placeholder="Name"
-              onChange={(event) => {
-                setUsername(event.target.value);
-              }}
-            />
-            <input
-              type="text"
-              placeholder="Room ID"
-              onChange={(event) => {
-                setRoom(event.target.value);
-              }}
-            />
-            <button onClick={joinRoom}>Join a Room</button>
-          </div>
-        ) : (
-
-            //Keep this part 
-            <Chat socket={socket} username={username} room={room} /> 
-            
-        )}
-      </div> */}
 
     </main>
